digital-shop-app/mainapp/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from mainapp.views import CustomTemplateView, UserDetail, EntityViewSet, UserViewSet
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView


router = DefaultRouter()
router.register('entities', EntityViewSet, basename='users')

urlpatterns = [
    #path('', CustomTemplateView.as_view(), name='home'),
    path('api/v1/', include(router.urls)),
    path('api-auth/', include('rest_framework.urls')),
    path('api/v1/rest-auth/', include('rest_auth.urls')),
    path('api/v1/rest-auth/registration/', include('rest_auth.registration.urls')),
    # JWT Token
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
